backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
asyncpg
alembic
passlib[argon2]
python-jose
pydantic
pydantic[email]